usecases TDD: Product and Retailer

+++ prisma/migrations/20231210052129_schema/migration.sql
@@ -19,7 +19,7 @@ CREATE TABLE "Affiliate" (
 CREATE TABLE "Product" (
     "id" UUID NOT NULL,
     "name" TEXT NOT NULL,
-    "affiliatePercent" INTEGER NOT NULL,
+    "affiliatePercent" DOUBLE PRECISION NOT NULL,
     "retailerId" UUID NOT NULL,
     "affiliateId" UUID,
